backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1710965072196/work
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
blinker @ file:///work/perseverance-python-buildout/croot/blinker_1698846408677/work
boltons @ file:///work/perseverance-python-buildout/croot/boltons_1698851177130/work
Brotli @ file:///croot/brotli-split_1714483155106/work
certifi==2024.8.30
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.3.2
click @ file:///work/perseverance-python-buildout/croot/click_1698845879718/work
conda @ file:///croot/conda_1722004606466/work
conda-content-trust @ file:///croot/conda-content-trust_1714483159009/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1706733287605/work/src
conda-package-handling @ file:///croot/conda-package-handling_1718138267740/work
conda_package_streaming @ file:///croot/conda-package-streaming_1718136078615/work
constantly==23.10.4
contourpy==1.3.0
cryptography==43.0.1
cycler==0.12.1
daphne==4.1.2
distro @ file:///croot/distro_1714488253808/work
Django==5.1.1
django-environ==0.11.2
django-redis==5.4.0
django-restframework==0.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Flask @ file:///croot/flask_1716545870149/work
fonttools==4.54.1
frozendict @ file:///croot/frozendict_1713194832637/work
frozenlist==1.4.1
git-filter-repo==2.45.0
gunicorn==21.2.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itsdangerous @ file:///croot/itsdangerous_1716533337253/work
Jinja2 @ file:///croot/jinja2_1716993405101/work
jsonpatch @ file:///croot/jsonpatch_1714483231291/work
jsonpointer==2.1
kiwisolver==1.4.7
libmambapy @ file:///croot/mamba-split_1714483352891/work/libmambapy
ma==1.0.2
ma.mongo==1.0.2
MarkupSafe @ file:///work/perseverance-python-buildout/croot/markupsafe_1707342994941/work
matplotlib==3.9.2
menuinst @ file:///croot/menuinst_1718132535249/work
msgpack==1.1.0
multidict==6.1.0
mutagen==1.47.0
numpy==2.1.1
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.3
paypalrestsdk==1.13.3
pillow==10.4.0
platformdirs @ file:///work/perseverance-python-buildout/croot/platformdirs_1701732573265/work
pluggy @ file:///work/perseverance-python-buildout/croot/pluggy_1698805497733/work
priority==1.3.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycosat @ file:///croot/pycosat_1714510623388/work
pycparser==2.22
pycryptodomex==3.20.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
pyparsing==3.1.4
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1698845478203/work
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2024.2
redis==5.0.8
requests==2.32.3
ruamel.yaml @ file:///work/perseverance-python-buildout/croot/ruamel.yaml_1698863605521/work
service-identity==24.1.0
setuptools==75.1.0
six==1.16.0
sqlparse==0.5.1
tqdm @ file:///croot/tqdm_1716395931952/work
truststore @ file:///work/perseverance-python-buildout/croot/truststore_1701735771625/work
twilio==9.3.1
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
websockets==13.0
Werkzeug @ file:///croot/werkzeug_1716533313107/work
wheel==0.44.0
yarl==1.11.1
yt-dlp==2024.8.6
zope.interface==7.0.3
zstandard @ file:///croot/zstandard_1714677652653/work
